NOISSUE - Unify group and clients implementations on things and users (#1793)
* unify groups repo implementation


* unify clients implementation


* closer client integration


* further unification of groups


* enable on update secret & owner


* unify retrieve all


* fully unify groups repository


* add secret to retrieve all


* save updated at


* fix test


* fix retrieve all tests


* restore files


* fix build


* remove unused files


* fix retrieve all tests


* fix linting


* fix linting


* restore broken changes


* restore setup tests


* update where condition


* remove extra db object


* unify groups test


* unify clients test


* remove unused variables


* update changes


* sync with master current updates


* update test


* fix tests


* fix test


* fix test


* fix tests


* match changes in #1877


* separate things and users repos


* remove comments
 implement retrieveBysecret in things only


* remove exec


* remove duplicate imports


* wrap errors


---------
