+++ pkg/clients/postgres/doc.go
@@ -1,0 +1,5 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+// Package postgres contains the database implementation of clients repository layer.
+package postgres
