+++ pkg/clients/postgres/setup_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/jmoiron/sqlx"
 	pgClient "github.com/mainflux/mainflux/internal/clients/postgres"
 	"github.com/mainflux/mainflux/internal/postgres"
-	gpostgres "github.com/mainflux/mainflux/things/postgres"
+	upostgres "github.com/mainflux/mainflux/users/postgres"
 	dockertest "github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
 	"go.opentelemetry.io/otel"
@@ -75,7 +75,8 @@ func TestMain(m *testing.M) {
 		SSLKey:      "",
 		SSLRootCert: "",
 	}
-	if db, err = pgClient.SetupDB(dbConfig, *gpostgres.Migration()); err != nil {
+
+	if db, err = pgClient.SetupDB(dbConfig, *upostgres.Migration()); err != nil {
 		log.Fatalf("Could not setup test DB connection: %s", err)
 	}
