--- users/groups/postgres/doc.go
@@ -1,5 +1,0 @@
-// Copyright (c) Mainflux
-// SPDX-License-Identifier: Apache-2.0
-
-// Package postgres contains the database implementation of groups repository layer.
-package postgres
