+++ users/policies/postgres/policies_test.go
@@ -12,9 +12,9 @@ import (
 	mfclients "github.com/mainflux/mainflux/pkg/clients"
 	"github.com/mainflux/mainflux/pkg/errors"
 	mfgroups "github.com/mainflux/mainflux/pkg/groups"
+	gpostgres "github.com/mainflux/mainflux/pkg/groups/postgres"
 	"github.com/mainflux/mainflux/pkg/uuid"
 	cpostgres "github.com/mainflux/mainflux/users/clients/postgres"
-	gpostgres "github.com/mainflux/mainflux/users/groups/postgres"
 	"github.com/mainflux/mainflux/users/policies"
 	ppostgres "github.com/mainflux/mainflux/users/policies/postgres"
 	"github.com/stretchr/testify/assert"
@@ -29,7 +29,7 @@ func TestPoliciesSave(t *testing.T) {
 	t.Cleanup(func() { testsutil.CleanUpDB(t, db) })
 	repo := ppostgres.NewRepository(database)
 	crepo := cpostgres.NewRepository(database)
-	grepo := gpostgres.NewRepository(database)
+	grepo := gpostgres.New(database)
 
 	group := mfgroups.Group{
 		ID:   testsutil.GenerateUUID(t, idProvider),
@@ -50,7 +50,7 @@ func TestPoliciesSave(t *testing.T) {
 
 	clients, err := crepo.Save(context.Background(), client)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
-	client = clients[0]
+	client = clients
 
 	cases := []struct {
 		desc   string
@@ -89,7 +89,7 @@ func TestPoliciesEvaluate(t *testing.T) {
 	t.Cleanup(func() { testsutil.CleanUpDB(t, db) })
 	repo := ppostgres.NewRepository(database)
 	crepo := cpostgres.NewRepository(database)
-	grepo := gpostgres.NewRepository(database)
+	grepo := gpostgres.New(database)
 
 	client1 := mfclients.Client{
 		ID:   testsutil.GenerateUUID(t, idProvider),
@@ -116,10 +116,10 @@ func TestPoliciesEvaluate(t *testing.T) {
 
 	clients1, err := crepo.Save(context.Background(), client1)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
-	client1 = clients1[0]
+	client1 = clients1
 	clients2, err := crepo.Save(context.Background(), client2)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
-	client2 = clients2[0]
+	client2 = clients2
 	group, err = grepo.Save(context.Background(), group)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
 
@@ -178,7 +178,7 @@ func TestPoliciesRetrieve(t *testing.T) {
 	t.Cleanup(func() { testsutil.CleanUpDB(t, db) })
 	repo := ppostgres.NewRepository(database)
 	crepo := cpostgres.NewRepository(database)
-	grepo := gpostgres.NewRepository(database)
+	grepo := gpostgres.New(database)
 
 	group := mfgroups.Group{
 		ID:   testsutil.GenerateUUID(t, idProvider),
@@ -199,7 +199,7 @@ func TestPoliciesRetrieve(t *testing.T) {
 
 	clients, err := crepo.Save(context.Background(), client)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
-	client = clients[0]
+	client = clients
 
 	policy := policies.Policy{
 		OwnerID: client.ID,
@@ -234,7 +234,7 @@ func TestPoliciesUpdate(t *testing.T) {
 	t.Cleanup(func() { testsutil.CleanUpDB(t, db) })
 	repo := ppostgres.NewRepository(database)
 	crepo := cpostgres.NewRepository(database)
-	grepo := gpostgres.NewRepository(database)
+	grepo := gpostgres.New(database)
 
 	group := mfgroups.Group{
 		ID:   testsutil.GenerateUUID(t, idProvider),
@@ -352,7 +352,7 @@ func TestPoliciesRetrievalAll(t *testing.T) {
 	t.Cleanup(func() { testsutil.CleanUpDB(t, db) })
 	repo := ppostgres.NewRepository(database)
 	crepo := cpostgres.NewRepository(database)
-	grepo := gpostgres.NewRepository(database)
+	grepo := gpostgres.New(database)
 
 	var nPolicies = uint64(10)
 
@@ -377,10 +377,10 @@ func TestPoliciesRetrievalAll(t *testing.T) {
 
 	clientsA, err := crepo.Save(context.Background(), clientA)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
-	clientA = clientsA[0]
+	clientA = clientsA
 	clientsB, err := crepo.Save(context.Background(), clientB)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
-	clientB = clientsB[0]
+	clientB = clientsB
 
 	grps := []string{}
 	for i := uint64(0); i < nPolicies; i++ {
@@ -628,7 +628,7 @@ func TestPoliciesDelete(t *testing.T) {
 	t.Cleanup(func() { testsutil.CleanUpDB(t, db) })
 	repo := ppostgres.NewRepository(database)
 	crepo := cpostgres.NewRepository(database)
-	grepo := gpostgres.NewRepository(database)
+	grepo := gpostgres.New(database)
 
 	group := mfgroups.Group{
 		ID:   testsutil.GenerateUUID(t, idProvider),
@@ -649,7 +649,7 @@ func TestPoliciesDelete(t *testing.T) {
 
 	clients, err := crepo.Save(context.Background(), client)
 	require.Nil(t, err, fmt.Sprintf("unexpected error: %s", err))
-	client = clients[0]
+	client = clients
 
 	policy := policies.Policy{
 		OwnerID: client.ID,
